refactor: settings dialog and related panels
